+++ Widex Compass GPS Checker.py
@@ -4,9 +4,9 @@
 #                                                           #
 #############################################################
 import os
-import html
-import ast
 import json
+import shutil
+import zipfile
 import requests
 import rot_codec
 from pathlib import Path
@@ -84,7 +84,7 @@ except:
 # Target market input
 possibleMarkets = ("Acuitis", "Albania", "Algeria", "Alternate_Test_Distributor", "Argentina", "Audigy", "Australia", "Austria", "Belarus", "Belgium", "Bolivia", "Bosnia Herzegovina", "Brazil", "Bulgaria", "Canada", "Chile", "China", "Colombia", "ConHearing", "Costa Rica", "Croatia", "Cyprus", "Czech Republic", "Denmark", "Dominican Republic", "Ecuador", "Egypt", "El Salvador", "Estonia", "F.Y.R.O.M.", "Finland", "France", "Germany", "Ghana", "Grand Audition", "Greece", "Guyana", "Hong Kong", "Huier", "Hungary", "India", "Indonesia", "Iran", "Iraq", "Ireland", "Israel", "Italy", "Jamaica", "Japan", "Jordan", "Kazakhstan", "Kenya", "Korea", "Kosovo", "Kuwait", "Kyrgyzstan", "Latvia", "Lebanon", "Libya", "Lithuania", "Main_Test_Distributor", "Malaysia", "Malta", "Mexico", "Mongolia", "Morocco", "Namibia", "Netherlands", "New Zealand", "Nigeria", "Norway", "Oman", "Pakistan", "Panama", "Paraguay", "Peru", "Philippines", "Poland", "Portugal", "Romania", "Russia", "Saudi Arabia", "SayWhat", "Serbia", "Singapore", "Slovakia", "Slovenia", "South Africa", "Spain", "Specsavers", "Sri Lanka", "Sudan", "Sweden", "Switzerland", "Syria", "Taiwan", "Thailand", "Tunisia", "Turkey", "Ukraine", "United Arab Emirates", "United Kingdom", "Uruguay", "USA", "Venezuela", "Veteran Affairs US", "Vietnam", "Widex Monitoring System", "Yemen")
 possibleMarketsSet = set(possibleMarkets)
-inputMarket = input("\nPlease enter " + Fore.GREEN + "target market ID (Case-Sensitive!)" + Style.RESET_ALL + " [" + defaultMarketSrc + "default: " + Fore.YELLOW + defaultMarket + Style.RESET_ALL + "]: ")
+inputMarket = input("\nPlease enter " + Fore.GREEN + "target Distributor ID (Case-Sensitive!)" + Style.RESET_ALL + " [" + defaultMarketSrc + "default: " + Fore.YELLOW + defaultMarket + Style.RESET_ALL + "]: ")
 if (inputMarket == "" or inputMarket.lower() == defaultMarket.lower()):
     targetMarket = defaultMarket
     print("\nChecking for " + Fore.GREEN + targetMarket + Style.RESET_ALL + " market...")
@@ -93,7 +93,7 @@ elif (inputMarket in possibleMarketsSet):
     print("\nChecking for " + Fore.GREEN + targetMarket + Style.RESET_ALL + " market...")
 else:
     targetMarket = defaultMarket
-    print("\n" + Fore.RED + "Error" + Style.RESET_ALL + ": Market ID is invalid. Using " + Fore.GREEN + defaultMarket + Style.RESET_ALL + " instead...")
+    print("\n" + Fore.RED + "Error" + Style.RESET_ALL + ": Distributor ID is invalid. Using " + Fore.GREEN + defaultMarket + Style.RESET_ALL + " instead...")
 
 # Base information
 baseId = "CompassGPS"
@@ -141,62 +141,66 @@ if "Packages" in data:
     if (data['Packages']==[]):
         print("\n\nThe latest available " + baseId + " version is " + Fore.GREEN + "v" + baseVer + Style.RESET_ALL + "\n\n")
         print("No further update available.")
+        exit(1)
     else:
-        appVer = data['Packages'][0]['Version'] + "." + str(data['Packages'][0]['Revision'])
-        print("\n\nThe latest available " + baseId + " version is " + Fore.GREEN + "v" + appVer + Style.RESET_ALL + "\n\n")
-
-# UNFINISHED!!!!
-# (These are copied from Starkey. Further modification required.)
-#
-# appVer = data['Update']['Title'] + ' (' + data['Update']['Version'] + ')'
-# print ("Version info:")
-# print(appVer)
-# print(data['Update']['Description'])
-# filesList = json.dumps(ast.literal_eval(str(data['Update']['Files'][0])))
-# fileData = json.loads(filesList)
-#
-# availableFiles = [] # List of available files
-# availableFiles.append( (os.path.basename(fileData['Url']), fileData['Url']) )
+        latestVersion = data['Packages'][0]['Version'] + "." + str(data['Packages'][0]['Revision'])
 
 
+print("\n\nThe latest available " + baseId + " version for " + Fore.GREEN + targetMarket + Style.RESET_ALL + " distributor is " + Fore.GREEN + "v" + latestVersion + Style.RESET_ALL + "\n\n")
+
+# List of versions
+validVersions = [
+    (latestVersion, 'The latest available ' + baseId + ' verion'),
+]
+
+# Select outputDir and targetVersion
+outputDir = libhearingdownloader.selectOutputFolder()
+targetVersion = validVersions[libhearingdownloader.selectFromList(validVersions)][0]
+print("\n\n")
+
+# Create download folder
+outputDir += "Widex Compass GPS " + targetVersion + "/"
+if (libhearingdownloader.verboseDebug):
+    print(outputDir)
+
+print ("Downloading directory index")
+
+packageResources = data['Packages'][0]['Resources']
+
+filesToDownload = {}
+for resourceFile in packageResources:
+    filesToDownload[resourceFile['Value'].split("/")[-1]] = resourceFile['Value']
+
+if (libhearingdownloader.verboseDebug):
+    print(filesToDownload)
+
+# Download and save the files
+print("Downloading " + str(len(filesToDownload.keys())) + " files\n")
+currentFile = 1
+for fileToDownload in filesToDownload.keys():
+    if (libhearingdownloader.verboseDebug):
+        print(filesToDownload[fileToDownload])
+        print(outputDir + fileToDownload)
+        print("Downloading " + fileToDownload + " (" + str(currentFile) + "/" + str(len(filesToDownload.keys())) + ")")
+    # Download file
+    libhearingdownloader.downloadFile(filesToDownload[fileToDownload], outputDir + fileToDownload, "Downloading " + fileToDownload + " (" + str(currentFile) + "/" + str(len(filesToDownload.keys())) + ")")
+
+    currentFile += 1
 
 
-# Nothing to download YET!
-# filesToDownload.append( (setupFileURL.split("/")[-1], setupFileURL) )
-# 
-# if (libhearingdownloader.verboseDebug):
-#     print(filesToDownload)
-# 
-# # Select outputDir and targetFile
-# outputDir = libhearingdownloader.selectOutputFolder()
-# 
-# # Create download folder
-# outputDir += "Widex Compass GPS/"
-# print("\n\n")
-# 
-# 
-# # Download and save the files
-# print("Downloading " + str(len(filesToDownload)) + " files\n")
-# currentFile = 1
-# for fileToDownload in filesToDownload:
-#     if (libhearingdownloader.verboseDebug):
-#         print(fileToDownload)
-# 
-#     # Download file
-#     libhearingdownloader.downloadFile(fileToDownload[1], outputDir + fileToDownload[0], "Downloading " + fileToDownload[1].split("/")[-1] + " (" + str(currentFile) + "/" + str(len(filesToDownload)) + ")")
-# 
-#     currentFile += 1
-# 
-# print("\n\n")
-# print("Creating installer from zip")
-# with zipfile.ZipFile(outputDir + "Setup.exe.zip", 'r') as zipFile:
-#     zipFile.extractall(outputDir)
-# 
-# os.makedirs(outputDir + "installations/")
-# 
-# for file in filesToDownload[:-1]:
-#     shutil.move(outputDir + file[0], outputDir + "installations/" + file[0])
-# 
-# os.remove(outputDir + filesToDownload[-1][0])
-# 
-# print("\n\nDownload Complete!")
+print("\n\n")
+print("Creating installer from zip")
+with zipfile.ZipFile(outputDir + "Setup.exe.zip", 'r') as zipFile:
+    zipFile.extractall(outputDir)
+
+os.makedirs(outputDir + "installations/")
+
+currentFile = 1
+for fileDownloaded in filesToDownload.keys():
+    if fileDownloaded != "Setup.exe.zip":
+        shutil.move(outputDir + fileDownloaded, outputDir + "installations/" + fileDownloaded)
+    else:
+        os.remove(outputDir + fileDownloaded)
+    currentFile += 1
+
+print("\n\nDownload Complete!")