+++ CHANGES.md
@@ -241,3 +241,6 @@
 ### v2025.05.13
 - Unitron: fix TrueFit checker for v5.8.0.29338 by changing target market
 - Phonak/Unitron/Hansaton/Oticon/Bernafon/Sonic/Philips/Signia/Rexton/Audio Service/A&M/Widex: LiveUpdate-ish default target market code with local fallback (Note: From next Release version on, target market updates will no longer be logged in Changelog anymore.)
+
+### Pre-release (work-in-progress)
+- Widex: fix checker for Compass GPS v4.9+ [close [issue #1](https://github.com/JediLin/Hearing-Aids-Fitting-Software-Update-Checkers/issues/1)]